refactor: Moved Pane into component directory
